+++ src/index.ts
@@ -2,7 +2,8 @@ import * as fhirR4 from './FHIR-R4/classes/models-r4';
 import * as IfhirR4 from './FHIR-R4/interfaces/IModel';
 import * as fhirR3 from './FHIR-R3';
 import * as dstu2 from './FHIR-DSTU2';
-import { ResourceUtilities } from './library/ResourceUtilities/ResourceUtilities';
-import { BundleUtilities } from './library/BundleUtilities/BundleUtilities';
-export { fhirR4, fhirR3, IfhirR4, dstu2, ResourceUtilities, BundleUtilities };
+import { ResourceUtils } from './library/ResourceUtils/ResourceUtils';
+import { BundleUtils } from './library/BundleUtils/BundleUtils';
+import { QueryBuilder } from './library/QueryBuilder/QueryBuilder';
+export { fhirR4, fhirR3, IfhirR4, dstu2, ResourceUtils, BundleUtils, QueryBuilder };