+++ src/library/constants.ts
@@ -1,0 +1,4 @@
+export enum SORT_ORDER {
+    ASCENDING,
+    DESCENDING
+}